Removed redundant logging package and moved everything to profile.go - Fixed log/logf bug in passing args - Added missing multiple profiling test
